move parser into own file
