+++ .circleci/config.yml
@@ -28,6 +28,7 @@ jobs:
           command: |
               source .venv/bin/activate
               cv latex cvcreator/templates/example.toml output.tex
+              python -m cvcreator latex cvcreator/templates/example.toml output.tex
       # - run:
       #     name: "Run tests"
       #     command: |
